Fixing up FRAT & gates

+++ src/bva.h
@@ -62,7 +62,7 @@ public:
 private:
     Solver* solver;
     OccSimplifier* simplifier;
-    vector<uint16_t>& seen;
+    vector<uint32_t>& seen;
     vector<uint8_t>& seen2;
 
     Stats runStats;
