+++ src/distillerlongwithimpl.h
@@ -161,7 +161,7 @@ class DistillerLongWithImpl {
         Solver* solver;
         vector<Lit> lits;
         vector<Lit> lits2;
-        vector<uint16_t>& seen;
+        vector<uint32_t>& seen;
         vector<uint8_t>& seen2;
 
         //Global status
